requirements.txt
openai-whisper==20231117
pydub==0.25.1
srt==3.5.3
